Initial commit. Simple blog functionality working.
Includes: Add, Edit and Delete posts; Add and Delete comments.

Add post can only be performed by a logged-in user.
Edit and Delete posts requires the original poster or an administrator.
Add comment can be done by a logged-in or anonymous user.
Delete comment requires the original commenter or an administrator.
Deleting an anonymous comment requires an administrator.